requirements.txt
anthropic>=0.34.0
tavily-python>=0.5.0
python-dotenv>=1.0.1
requests>=2.32.0
openai>=2.6.0
httpx>=0.27.0

